+++ pkg/model/database/database.go
@@ -47,7 +47,7 @@ var ProviderFields = struct {
 // used to authenticate.
 type Provider struct {
 	ProviderID uint   `gorm:"primaryKey"`
-	Name       string `gorm:"size:20;not null"enum:"azuredevops,gitlab,github"`
+	Name       string `gorm:"size:20;not null" enum:"azuredevops,gitlab,github"`
 	URL        string `gorm:"size:500;not null"`
 	UploadURL  string `gorm:"size:500"`
 	TokenID    uint   `gorm:"nullable;default:NULL;index:provider_idx_token_id"`
@@ -129,7 +129,7 @@ var BranchFields = struct {
 // BranchColumns holds the DB column names for each field.
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
-var BranchColumns = struct{
+var BranchColumns = struct {
 	BranchID string
 }{
 	BranchID: "branch_id",
@@ -151,9 +151,11 @@ type Branch struct {
 // Useful in GORM .Where() statements to only select certain fields or in GORM
 // Preload statements to select the correct field to preload.
 var BuildFields = struct {
-	Params string
+	Params              string
+	TestResultSummaries string
 }{
-	Params: "Params",
+	Params:              "Params",
+	TestResultSummaries: "TestResultSummaries",
 }
 
 // BuildColumns holds the DB column names for each field.
@@ -190,7 +192,7 @@ type Build struct {
 	StartedOn           null.Time           `gorm:"nullable;default:NULL"`
 	CompletedOn         null.Time           `gorm:"nullable;default:NULL"`
 	GitBranch           string              `gorm:"size:300;default:'';not null"`
-	Environment         null.String         `gorm:"nullable;size:40"swaggertype:"string"`
+	Environment         null.String         `gorm:"nullable;size:40" swaggertype:"string"`
 	Stage               string              `gorm:"size:40;default:'';not null"`
 	Params              []BuildParam        `gorm:"foreignKey:BuildID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE"`
 	IsInvalid           bool                `gorm:"not null;default:false"`